# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnomalyRecordRisk = typeof AnomalyRecordRisk[keyof typeof AnomalyRecordRisk];


export const AnomalyRecordRisk = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface AnomalyRecord {
  id: string;
  recordId: string;
  enumeratorId: string;
  region: string;
  age: number;
  income: number;
  education?: string;
  status: string;
  risk: AnomalyRecordRisk;
  reason: string;
  detectedAt: string;
  score?: number;
}

export interface RegionalAnomaly {
  region: string;
  anomalies: number;
  total: number;
}

export interface IngestionPoint {
  label: string;
  records: number;
}

export type ActivityItemTone = typeof ActivityItemTone[keyof typeof ActivityItemTone];


export const ActivityItemTone = {
  blue: 'blue',
  green: 'green',
  orange: 'orange',
  red: 'red',
} as const;

export interface ActivityItem {
  id: string;
  title: string;
  detail: string;
  time: string;
  tone: ActivityItemTone;
}

export interface ValidationSummary {
  totalRecords: number;
  totalAnomalies: number;
  highRiskEnumerators: number;
  anomalyRate: number;
  regionalAnomalies: RegionalAnomaly[];
  ingestionTrend: IngestionPoint[];
  activity: ActivityItem[];
}

export interface UploadSurveyInput {
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  csvContent: string;
}

export interface UploadResult {
  fileName: string;
  recordsProcessed: number;
  anomaliesFound: number;
  highRiskCount: number;
}

export type ValidationRuleStatus = typeof ValidationRuleStatus[keyof typeof ValidationRuleStatus];


export const ValidationRuleStatus = {
  Active: 'Active',
  Draft: 'Draft',
} as const;

export interface ValidationRule {
  id: string;
  name: string;
  condition: string;
  action: string;
  status: ValidationRuleStatus;
  createdAt: string;
}

export interface ValidationRuleInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  condition: string;
  /** @minLength 1 */
  action: string;
}

export type GetAnomaliesParams = {
risk?: GetAnomaliesRisk;
search?: string;
};

export type GetAnomaliesRisk = typeof GetAnomaliesRisk[keyof typeof GetAnomaliesRisk];


export const GetAnomaliesRisk = {
  All: 'All',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;