# lib/api-zod/src/generated/types/anomalyRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnomalyRecordRisk } from './anomalyRecordRisk';

export interface AnomalyRecord {
  id: string;
  recordId: string;
  enumeratorId: string;
  region: string;
  age: number;
  income: number;
  education?: string;
  status: string;
  risk: AnomalyRecordRisk;
  reason: string;
  detectedAt: string;
  score?: number;
}