# lib/api-zod/src/generated/types/regionalAnomaly.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegionalAnomaly {
  region: string;
  anomalies: number;
  total: number;
}