# lib/api-zod/src/generated/types/uploadSurveyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadSurveyInput {
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  csvContent: string;
}