# lib/api-zod/src/generated/types/validationRule.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationRuleStatus } from './validationRuleStatus';

export interface ValidationRule {
  id: string;
  name: string;
  condition: string;
  action: string;
  status: ValidationRuleStatus;
  createdAt: string;
}