# lib/api-zod/src/generated/types/validationRuleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ValidationRuleInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  condition: string;
  /** @minLength 1 */
  action: string;
}