# lib/api-zod/src/generated/types/validationSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';
import type { IngestionPoint } from './ingestionPoint';
import type { RegionalAnomaly } from './regionalAnomaly';

export interface ValidationSummary {
  totalRecords: number;
  totalAnomalies: number;
  highRiskEnumerators: number;
  anomalyRate: number;
  regionalAnomalies: RegionalAnomaly[];
  ingestionTrend: IngestionPoint[];
  activity: ActivityItem[];
}